+++ lib/data_services_api/dataset.rb
@@ -10,12 +10,12 @@ module DataServicesApi
       @service = service
     end
 
-    def id
+    def id # steep:ignore MethodBodyTypeMismatch
       @json['@id']
     end
 
     def method_missing(attribute)
-      respond_to_missing?(attribute) ? @json[attribute.to_s] : super
+      respond_to_missing?(attribute, nil) ? @json[attribute.to_s] : super
     end
 
     def respond_to_missing?(attribute, _args = nil)
@@ -26,16 +26,12 @@ module DataServicesApi
       @json['data-api']
     end
 
-    def dataset
+    def dataset # steep:ignore MethodBodyTypeMismatch
       @json['dataset']
     end
 
     def structure_api
       @json['structure-api']
-    end
-
-    def describe_api
-      @json['describe-api']
     end
 
     def explain_api
@@ -44,23 +40,11 @@ module DataServicesApi
       structure_api.gsub(/structure$/, 'explain')
     end
 
-    def structure
-      return @structure if defined?(@structure)
-
-      description = service.api_get_json(structure_api)
-      aspects = description['aspects']
-      @structure = aspects.map { |json| Aspect.new(json, service) }
-    end
-
     def query(query)
       sapi_query_params = SapiNTConverter.new(query.to_json).to_sapint_query
       sapint_response = service.api_get_json(data_api, sapi_query_params)
       json_mode_compact = query.terms['@json_mode'] == 'compact'
       DSAPIResponseConverter.new(sapint_response, dataset, json_mode_compact).to_dsapi_response
-    end
-
-    def describe(uri)
-      service.api_get_json(describe_api, uri: uri)
     end
 
     def explain(query)