+++ test/data_services_api/data_services_api_test.rb
@@ -9,10 +9,8 @@ describe 'DataServicesAPI', 'the data services API' do
 
   it 'should be constructable with a given URL' do
     mock_notifier = MockNotifications.new
-    mock_logger = mock('logger')
-    mock_logger.expects(:info).at_least(0)
 
-    dsapi = DataServicesApi::Service.new(url: 'foo/bar', instrumenter: mock_notifier, logger: mock_logger)
+    dsapi = DataServicesApi::Service.new(url: 'foo/bar', instrumenter: mock_notifier)
 
     _(dsapi.url).must_match(%r{foo/bar})
   end