+++ test/data_services_api/query_generator_test.rb
@@ -24,6 +24,7 @@ describe 'DataServiceApi::QueryGenerator' do
   it 'should start with an empty query' do
     pattern = {}
     query = DataServicesApi::QueryGenerator.new
+
     _(query.to_json).wont_be_nil
     assert_query(pattern, query.to_json)
   end
@@ -49,6 +50,7 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should allow sorting to specified' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query({ '@sort' => [{ '@up' => 'foo' }] }, query.sort(:up, 'foo').to_json)
     assert_query({ '@sort' => [{ '@up' => 'foo' }, { '@down' => 'bar' }] },
                  query.sort(:up, 'foo').sort(:down, 'bar').to_json)
@@ -65,6 +67,7 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should allow an arbitrary relational operator to be added' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query(conjunction([{ foo: { '@le' => 1000 } }, { foo: { '@ge' => 100 } }]),
                  query.op(:le, :foo, 1000)
                       .op('@ge', :foo, 100)
@@ -80,6 +83,7 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should add a value type when required' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query(conjunction(foo: { '@ge' => { '@value' => '2014-01-01', '@type' => 'xsd:date' } }),
                  query.op(:ge, :foo, Date.parse('2014-01-01'))
                       .to_json)
@@ -87,6 +91,7 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should allow a text search option to be added' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query(conjunction('@search' => 'foo'),
                  query.search('foo')
                       .to_json)
@@ -94,6 +99,7 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should allow a text search against a specific property to be added' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query(conjunction('@search' => { '@value' => 'foo', '@property' => 'foo:bar' }),
                  query.search_property('foo:bar', 'foo')
                       .to_json)
@@ -101,6 +107,7 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should allow a text search against a specific aspect to be added' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query(conjunction('foo:aspect' => { '@search' => 'foo' }),
                  query.search_aspect('foo:aspect', 'foo')
                       .to_json)
@@ -108,6 +115,7 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should allow a text search against a specific property of an aspect to be added' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query(conjunction('foo:aspect' => { '@search' => { '@value' => 'foo', '@property' => 'foo:bar' } }),
                  query.search_aspect_property('foo:aspect', 'foo:bar', 'foo')
                       .to_json)
@@ -115,6 +123,7 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should allow the limit to be set on a search query' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query(conjunction('foo:aspect' => { '@search' => { '@value' => 'foo',
                                                               '@property' => 'foo:bar',
                                                               '@limit' => 997 } }),
@@ -129,6 +138,7 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should allow a simple boolean expression to be added' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query(conjunction('foo:aspect' => { '@oneof' => [{ '@id' => 'foo:bar' }, { '@id' => 'foo:bam' }] }),
                  query.eq_any_uri('foo:aspect', %w[foo:bar foo:bam])
                       .to_json)
@@ -140,6 +150,7 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should allow a type to be specified for a boolean expression value' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query(conjunction('foo:aspect' => { '@oneof' => [{ '@value' => 'foo:bar', '@type' => 'xsd:coconut' },
                                                             { '@value' => 'foo:bam', '@type' => 'xsd:coconut' }] }),
                  query.eq_any_value('foo:aspect', %w[foo:bar foo:bam], type: 'xsd:coconut')
@@ -148,6 +159,7 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should allow a terms to be composed' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query(conjunction([
                                { foo: { '@eq' => 'bar' } },
                                { cat: { '@eq' => 'cow' } },
@@ -159,6 +171,7 @@ describe 'DataServiceApi::QueryGenerator' do
                       .to_json)
 
     query = DataServicesApi::QueryGenerator.new
+
     assert_query(conjunction([
                                { 'foo:aspect' => { '@search' => { '@value' => 'foo', '@property' => 'foo:bar' } } },
                                { 'foo:aspect' => { '@search' => { '@value' => 'fim', '@property' => 'foo:blom' } } }
@@ -170,11 +183,13 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should allow a regex match to be specified' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query(conjunction('foo:aspect' => { '@matches' => 'bing.*' }),
                  query.matches('foo:aspect', 'bing.*')
                       .to_json)
 
     query = DataServicesApi::QueryGenerator.new
+
     assert_query(conjunction('foo:aspect' => { '@matches' => ['bing.*', 'i'] }),
                  query.matches('foo:aspect', 'bing.*', flags: 'i')
                       .to_json)
@@ -182,6 +197,7 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should allow an overall query limit to be set' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query({ '@limit' => 101, '@offset' => 20 },
                  query.limit(101)
                       .offset(20)
@@ -190,6 +206,7 @@ describe 'DataServiceApi::QueryGenerator' do
 
   it 'should allow a query to be run in count mode' do
     query = DataServicesApi::QueryGenerator.new
+
     assert_query({ '@count' => true }, query.count_only.to_json)
   end