+++ test/data_services_api/service_test.rb
@@ -11,18 +11,7 @@ class MockNotifications
 
   def instrument(*args)
     @instrumentations << args
-  end
-end
-
-class MockLogger
-  attr_reader :messages
-
-  def initialize
-    @messages = Hash.new { |h, k| h[k] = [] }
-  end
-
-  def info(message, &block)
-    @messages[:info] << [message, block&.call]
+    yield if block_given?
   end
 end
 
@@ -31,14 +20,10 @@ describe 'DataServicesAPI::Service' do
     ENV.fetch('API_SERVICE_URL', 'http://localhost:8888')
   end
 
-  let :mock_logger do
-    MockLogger.new
-  end
-
   before do
     mock_notifier = MockNotifications.new
     VCR.insert_cassette(name, record: :new_episodes)
-    @service = DataServicesApi::Service.new(url: api_url, instrumenter: mock_notifier, logger: mock_logger)
+    @service = DataServicesApi::Service.new(url: api_url, instrumenter: mock_notifier)
   end
 
   after do
@@ -47,12 +32,14 @@ describe 'DataServicesAPI::Service' do
 
   it 'should return the service URL' do
     mock_notifier = MockNotifications.new
-    service = DataServicesApi::Service.new(url: 'https://wimbledon.com', instrumenter: mock_notifier, logger: mock_logger)
+    service = DataServicesApi::Service.new(url: 'https://wimbledon.com', instrumenter: mock_notifier)
+
     _(service.url).must_equal('https://wimbledon.com')
   end
 
   it 'should find a dataset by name' do
     dataset = @service.dataset('ukhpi')
+
     _(dataset.data_api).must_match %r{/landregistry/id/ukhpi}
   end
 
@@ -65,8 +52,9 @@ describe 'DataServicesAPI::Service' do
   it 'should retrieve JSON with HTTP GET' do
     mock_notifier = MockNotifications.new
 
-    service = DataServicesApi::Service.new(url: api_url, instrumenter: mock_notifier, logger: mock_logger)
+    service = DataServicesApi::Service.new(url: api_url, instrumenter: mock_notifier)
     json = service.api_get_json("#{api_url}/landregistry/id/ukhpi", { '_limit' => 1 })
+
     _(json).wont_be_nil
     _(json['meta']).wont_be_nil
   end
@@ -75,78 +63,116 @@ describe 'DataServicesAPI::Service' do
     mock_notifier = MockNotifications.new
 
     DataServicesApi::Service
-      .new(url: api_url, instrumenter: mock_notifier, logger: mock_logger)
+      .new(url: api_url, instrumenter: mock_notifier)
       .api_get_json("#{api_url}/landregistry/id/ukhpi", { '_limit' => 1 })
 
-    instrumentations = mock_notifier.instrumentations
-    _(instrumentations.size).must_equal 1
-    _(instrumentations.first.first).must_equal 'response.api'
+    event_names = mock_notifier.instrumentations.map(&:first)
+
+    _(event_names).must_include 'response.data_services_api'
   end
 
   it 'should instrument a failed API call' do
     mock_api_url = 'http://localhost:8765'
     mock_notifier = MockNotifications.new
-    mock_logger = MockLogger.new
 
     _ do
       DataServicesApi::Service
-        .new(url: mock_api_url, instrumenter: mock_notifier, logger: mock_logger)
+        .new(url: mock_api_url, instrumenter: mock_notifier)
         .api_get_json("#{mock_api_url}/landregistry/id/ukhpi", { '_limit' => 1 })
     end.must_raise
 
-    instrumentations = mock_notifier.instrumentations
-    _(instrumentations.size).must_equal 1
-    _(instrumentations.first.first).must_equal 'connection_failure.api'
+    event_names = mock_notifier.instrumentations.map(&:first)
+
+    _(event_names).must_include 'connection_failure.data_services_api'
   end
 
   it 'should also instrument an API Service Exception' do
     mock_notifier = MockNotifications.new
-    mock_logger = MockLogger.new
 
-    _ do
+    error = _ do
       DataServicesApi::Service
-        .new(url: api_url, instrumenter: mock_notifier, logger: mock_logger)
+        .new(url: api_url, instrumenter: mock_notifier)
         .api_get_json("#{api_url}/ceci/nest/pas/une/page", { '_limit' => 1 })
-    end.must_raise
+    end.must_raise DataServicesApi::ServiceException
 
-    instrumentations = mock_notifier.instrumentations
-    _(instrumentations.size).must_equal 1
-    _(instrumentations.first.first).must_equal 'service_exception.api'
+    _(error.status).must_equal 404
+
+    _, payload = mock_notifier.instrumentations.find { |n, _| n == 'service_exception.data_services_api' }
+
+    _(payload).wont_be_nil
+    _(payload[:status]).must_equal 404
+    _(payload[:query_string]).must_equal '_limit=1'
   end
 
-  it 'should log the call to the data API' do
+  it 'should let subscribers derive the returned row count from the response payload' do
     mock_notifier = MockNotifications.new
-    mock_logger = MockLogger.new
 
     DataServicesApi::Service
-      .new(url: api_url, instrumenter: mock_notifier, logger: mock_logger)
+      .new(url: api_url, instrumenter: mock_notifier)
       .api_get_json("#{api_url}/landregistry/id/ukhpi", { '_limit' => 1 })
 
-    # @TODO: add specific constraints on received log messages
-    _(mock_logger.messages).wont_be_empty
+    _, payload = mock_notifier.instrumentations.find { |name, _| name == 'response.data_services_api' }
+
+    _(payload).wont_be_nil
+    _(payload[:response].body['items']).wont_be_nil
   end
 
   it 'should correctly receive a duration in microseconds' do
     mock_notifier = MockNotifications.new
-    mock_logger = MockLogger.new
 
     DataServicesApi::Service
-      .new(url: api_url, instrumenter: mock_notifier, logger: mock_logger)
+      .new(url: api_url, instrumenter: mock_notifier)
       .api_get_json("#{api_url}/landregistry/id/ukhpi", { '_limit' => 1 })
 
-    _(mock_logger).wont_be_nil
+    _, payload = mock_notifier.instrumentations.find { |name, _| name == 'response.data_services_api' }
 
-    # Check the last logged message for duration
-    mock_log = mock_logger.messages[:info].last
-    _(mock_log).wont_be_nil
-    _(mock_log.size).must_equal 2
+    _(payload).wont_be_nil
+    _(payload[:duration]).must_be :>, 0
+  end
 
-    json = mock_log.first
-    duration = JSON.parse(json)['request_time']
+  it 'should instrument a request before it is sent' do
+    mock_notifier = MockNotifications.new
 
-    if duration
-      _(duration.to_f).must_be :>, 0
-      assert_kind_of(String, duration)
-    end
+    DataServicesApi::Service
+      .new(url: api_url, instrumenter: mock_notifier)
+      .api_get_json("#{api_url}/landregistry/id/ukhpi", { '_limit' => 1 })
+
+    _, payload = mock_notifier.instrumentations.find { |name, _| name == 'request.data_services_api' }
+
+    _(payload).wont_be_nil
+    _(payload[:path]).must_equal '/landregistry/id/ukhpi'
+    _(payload[:query_string]).must_equal '_limit=1'
+  end
+
+  it 'should instrument a request even when the connection subsequently fails' do
+    mock_api_url = 'http://localhost:8765'
+    mock_notifier = MockNotifications.new
+
+    _ do
+      DataServicesApi::Service
+        .new(url: mock_api_url, instrumenter: mock_notifier)
+        .api_get_json("#{mock_api_url}/landregistry/id/ukhpi", { '_limit' => 1 })
+    end.must_raise
+
+    event_names = mock_notifier.instrumentations.map(&:first)
+
+    _(event_names).must_include 'request.data_services_api'
+  end
+
+  it 'should instrument each retry attempt before giving up on a failed connection' do
+    mock_api_url = 'http://localhost:8765'
+    mock_notifier = MockNotifications.new
+
+    _ do
+      DataServicesApi::Service
+        .new(url: mock_api_url, instrumenter: mock_notifier,
+             connection_failed_retry_options: { max: 2, interval: 0 })
+        .api_get_json("#{mock_api_url}/landregistry/id/ukhpi", { '_limit' => 1 })
+    end.must_raise
+
+    retries = mock_notifier.instrumentations.select { |entry| entry.first == 'retry.data_services_api' }
+
+    _(retries.size).must_equal 2
+    _(retries.map { |_, payload| payload[:retry_count] }).must_equal [1, 2]
   end
 end